create update clients command
